+++ rabin-karp.h
@@ -23,9 +23,9 @@ void precompute_RM();
 
 int pattern_match(unsigned long long int rkhash);
 
-unsigned long long int Rabin_Karp_Hash(char substring[],unsigned long long int start_index,unsigned long long int end_index,int newchunk, unsigned long long int hash_prev);
+unsigned long long int Rabin_Karp_Hash(char substring[], unsigned long long int start_index, unsigned long long int end_index, int newchunk, unsigned long long int hash_prev);
 
-char * copy_substring(char *str, char *s, unsigned long long int start,unsigned long long int end);
+char * copy_substring(char *str, char *s, off_t start, off_t end);
 
 void create_chunkfile(char str[],char shastr[], size_t);
